refactor: restore commented staging check in get-vehicles
- Restore commented staging environment check
- Keep staging fallback logic for future use
- Maintain code structure for different environments
