+++ src/app/(app)/pessoas/cnpj/[cnpj]/page.tsx
@@ -71,7 +71,6 @@ export default function Pessoa({ params: { cnpj } }: PessoaProps) {
   const remainingCredits = cortexRemainingCredits?.remaining_credit
   const timeUntilReset = cortexRemainingCredits?.time_until_reset
 
-
   useEffect(() => {
     async function initializeMap() {
       if (!mapContainerRef.current || !data) return
