+++ src/app/(app)/veiculos/busca-por-radar/components/selected-radars-list.tsx
@@ -1,38 +1,10 @@
-import { NavigationIcon, XCircleIcon } from 'lucide-react'
 import { memo, useCallback, useMemo, useRef } from 'react'
 
 import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
 import type { Radar } from '@/models/entities'
 
-interface RadarItemProps {
-  radar: Radar
-  onFocus: (radar: Radar) => void
-  onRemove: (radar: Radar) => void
-}
-
-const RadarItem = memo(function RadarItem({
-  radar,
-  onFocus,
-  onRemove,
-}: RadarItemProps) {
-  return (
-    <div className="flex items-center justify-between rounded bg-secondary p-2">
-      <div>
-        <div className="font-medium">{radar.cetRioCode}</div>
-        <div className="text-sm text-muted-foreground">{radar.location}</div>
-      </div>
-      <div className="flex space-x-2">
-        <Button variant="ghost" size="sm" onClick={() => onFocus(radar)}>
-          <NavigationIcon className="size-4" />
-        </Button>
-        <Button variant="ghost" size="sm" onClick={() => onRemove(radar)}>
-          <XCircleIcon className="size-4" />
-        </Button>
-      </div>
-    </div>
-  )
-})
+import { RadarItem } from './radar-item'
 
 interface SelectedRadarsListProps {
   selectedObjects: Radar[]
